Adding weapons package
Includes the unused attributes file that could be useful in the future. Includes the ALWSwordItem file which is mostly a copy paste of SwordItem.java file and AttributeUUID file for increased reach on weapons. ZweihanderItem.java is for the Zweihander weapons with the reach increase and I might update it later to have a set attack speed as well, not as necessary as the reach part though.